ai/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic-settings
python-multipart
supabase
postgrest
urllib3<2.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

llama-index>=0.10.0
llama-index-llms-openai>=0.1.0
llama-index-embeddings-openai>=0.1.0
llama-index-vector-stores-postgres>=0.1.0
llama-parse>=0.4.0
openai>=1.0.0
psycopg2-binary>=2.9.0